Refactor pcapng: move NameResolution to its own file
