+++ src/pcapng.rs
@@ -38,10 +38,10 @@ use crate::endianness::*;
 use crate::error::PcapError;
 use crate::traits::*;
 use crate::utils::*;
-use nom::bytes::streaming::{tag, take};
+use nom::bytes::streaming::take;
 use nom::combinator::{complete, map, map_parser};
 use nom::error::*;
-use nom::multi::{many0, many1, many_till};
+use nom::multi::{many0, many1};
 use nom::number::streaming::{be_u32, le_u32};
 use nom::{Err, IResult};
 use rusticata_macros::{align32, newtype_enum};
@@ -50,11 +50,13 @@ use std::convert::TryFrom;
 
 mod enhanced_packet;
 mod interface_description;
+mod name_resolution;
 mod section_header;
 mod simple_packet;
 
 pub use enhanced_packet::*;
 pub use interface_description::*;
+pub use name_resolution::*;
 pub use section_header::*;
 pub use simple_packet::*;
 
@@ -281,72 +283,6 @@ pub fn build_ts_f64(ts_high: u32, ts_low: u32, ts_offset: u64, resolution: u64) 
     let ts_fractional = (ts % resolution) as u32;
     // XXX should we round to closest unit?
     ts_sec as f64 + ((ts_fractional as f64) / (resolution as f64))
-}
-
-#[derive(Clone, Copy, Eq, PartialEq)]
-pub struct NameRecordType(pub u16);
-
-newtype_enum! {
-    impl debug NameRecordType {
-        End = 0,
-        Ipv4 = 1,
-        Ipv6 = 2
-    }
-}
-
-#[derive(Debug)]
-pub struct NameRecord<'a> {
-    pub record_type: NameRecordType,
-    pub record_value: &'a [u8],
-}
-
-impl<'a> NameRecord<'a> {
-    pub const END: NameRecord<'static> = NameRecord {
-        record_type: NameRecordType::End,
-        record_value: &[],
-    };
-}
-
-#[derive(Debug)]
-pub struct NameResolutionBlock<'a> {
-    pub block_type: u32,
-    pub block_len1: u32,
-    pub nr: Vec<NameRecord<'a>>,
-    pub options: Vec<PcapNGOption<'a>>,
-    pub block_len2: u32,
-}
-
-impl<'a, En: PcapEndianness> PcapNGBlockParser<'a, En, NameResolutionBlock<'a>>
-    for NameResolutionBlock<'a>
-{
-    const HDR_SZ: usize = 12;
-    const MAGIC: u32 = NRB_MAGIC;
-
-    fn inner_parse<E: ParseError<&'a [u8]>>(
-        block_type: u32,
-        block_len1: u32,
-        i: &'a [u8],
-        block_len2: u32,
-    ) -> IResult<&'a [u8], NameResolutionBlock<'a>, E> {
-        let start_i = i;
-        // caller function already tested header type(magic) and length
-        // read records
-        let (i, nr) = parse_name_record_list::<En, E>(i)?;
-        // read options
-        let current_offset = 12 + (i.as_ptr() as usize) - (start_i.as_ptr() as usize);
-        let (i, options) = opt_parse_options::<En, E>(i, block_len1 as usize, current_offset)?;
-        if block_len2 != block_len1 {
-            return Err(Err::Error(E::from_error_kind(i, ErrorKind::Verify)));
-        }
-        let block = NameResolutionBlock {
-            block_type,
-            block_len1,
-            nr,
-            options,
-            block_len2,
-        };
-        Ok((i, block))
-    }
 }
 
 #[derive(Debug)]
@@ -808,46 +744,6 @@ fn if_extract_tsoffset_and_tsresol(options: &[PcapNGOption]) -> (u8, i64) {
         }
     }
     (if_tsresol, if_tsoffset)
-}
-
-fn parse_name_record<'a, En: PcapEndianness, E: ParseError<&'a [u8]>>(
-    i: &'a [u8],
-) -> IResult<&'a [u8], NameRecord, E> {
-    let (i, record_type) = En::parse_u16(i)?;
-    let (i, record_len) = En::parse_u16(i)?;
-    let aligned_len = align32!(record_len as u32);
-    let (i, record_value) = take(aligned_len)(i)?;
-    let name_record = NameRecord {
-        record_type: NameRecordType(record_type),
-        record_value,
-    };
-    Ok((i, name_record))
-}
-
-fn parse_name_record_list<'a, En: PcapEndianness, E: ParseError<&'a [u8]>>(
-    i: &'a [u8],
-) -> IResult<&'a [u8], Vec<NameRecord>, E> {
-    map(
-        many_till(parse_name_record::<En, E>, tag(b"\x00\x00\x00\x00")),
-        |(mut v, _)| {
-            v.push(NameRecord::END);
-            v
-        },
-    )(i)
-}
-
-#[inline]
-pub fn parse_nameresolutionblock_le(
-    i: &[u8],
-) -> IResult<&[u8], NameResolutionBlock, PcapError<&[u8]>> {
-    ng_block_parser::<NameResolutionBlock, PcapLE, _, _>()(i)
-}
-
-#[inline]
-pub fn parse_nameresolutionblock_be(
-    i: &[u8],
-) -> IResult<&[u8], NameResolutionBlock, PcapError<&[u8]>> {
-    ng_block_parser::<NameResolutionBlock, PcapBE, _, _>()(i)
 }
 
 pub fn parse_interfacestatisticsblock_le(
